Create HTMl-using Links
This HTML page demonstrates different types of hyperlinks. It includes a normal text link to the AEC website, a link that opens in a new tab using the target="_blank" attribute, and an image used as a clickable link that opens the AEC gallery. The page also uses headings and horizontal lines to organize the content.